Struct cw_storage_plus::IndexedMap

pub struct IndexedMap<'a, K, T, I>where
    K: PrimaryKey<'a>,
    T: Serialize + DeserializeOwned + Clone,
    I: IndexList<T>,{
    pub idx: I,
    /* private fields */
}

IndexedMap works like a Map but has a secondary index

Fields
idx: I

This is meant to be read directly to get the proper types, like: map.idx.owner.items(…)

Implementations

impl<'a, K, T, I> IndexedMap<'a, K, T, I>where
K: PrimaryKey<'a>,
T: Serialize + DeserializeOwned + Clone,
I: IndexList<T>,

pub fn new(pk_namespace: &'a str, indexes: I) -> Self

pub fn key(&self, k: K) -> Path<T>

impl<'a, K, T, I> IndexedMap<'a, K, T, I>where
K: PrimaryKey<'a>,
T: Serialize + DeserializeOwned + Clone,
I: IndexList<T>,

pub fn save(&self, store: &mut dyn Storage, key: K, data: &T) -> StdResult<()>

save will serialize the model and store, returns an error on serialization issues. this must load the old value to update the indexes properly if you loaded the old value earlier in the same function, use replace to avoid needless db reads

pub fn remove(&self, store: &mut dyn Storage, key: K) -> StdResult<()>

pub fn replace(
&self,
store: &mut dyn Storage,
key: K,
data: Option<&T>,
old_data: Option<&T>
) -> StdResult<()>

replace writes data to key. old_data must be the current stored value (from a previous load) and is used to properly update the index. This is used by save, replace, and update and can be called directly if you want to optimize

pub fn update<A, E>(
&self,
store: &mut dyn Storage,
key: K,
action: A
) -> Result<T, E>where
A: FnOnce(Option<T>) -> Result<T, E>,
E: From<StdError>,

Loads the data, perform the specified action, and store the result in the database. This is shorthand for some common sequences, which may be useful.

If the data exists, action(Some(value)) is called. Otherwise action(None) is called.

pub fn load(&self, store: &dyn Storage, key: K) -> StdResult<T>

load will return an error if no data is set at the given key, or on parse error

pub fn may_load(&self, store: &dyn Storage, key: K) -> StdResult<Option<T>>

may_load will parse the data stored at the key if present, returns Ok(None) if no data there. returns an error on issues parsing

pub fn has(&self, store: &dyn Storage, k: K) -> bool

Returns true if storage contains this key, without parsing or interpreting the contents.

impl<'a, K, T, I> IndexedMap<'a, K, T, I>where
K: PrimaryKey<'a>,
T: Serialize + DeserializeOwned + Clone,
I: IndexList<T>,

pub fn prefix_range_raw<'c>(
&self,
store: &'c dyn Storage,
min: Option<PrefixBound<'a, K::Prefix>>,
max: Option<PrefixBound<'a, K::Prefix>>,
order: Order
) -> Box<dyn Iterator<Item = StdResult<Record<T>>> + 'c>where
T: 'c,
'a: 'c,

While range_raw over a prefix fixes the prefix to one element and iterates over the remaining, prefix_range_raw accepts bounds for the lowest and highest elements of the Prefix itself, and iterates over those (inclusively or exclusively, depending on PrefixBound). There are some issues that distinguish these two, and blindly casting to Vec<u8> doesn’t solve them.

impl<'a, K, T, I> IndexedMap<'a, K, T, I>where
T: Serialize + DeserializeOwned + Clone,
K: PrimaryKey<'a>,
I: IndexList<T>,

pub fn sub_prefix(
&self,
p: K::SubPrefix
) -> Prefix<K::SuperSuffix, T, K::SuperSuffix>

pub fn prefix(&self, p: K::Prefix) -> Prefix<K::Suffix, T, K::Suffix>

impl<'a, K, T, I> IndexedMap<'a, K, T, I>where
T: Serialize + DeserializeOwned + Clone,
K: PrimaryKey<'a> + KeyDeserialize,
I: IndexList<T>,

pub fn prefix_range<'c>(
&self,
store: &'c dyn Storage,
min: Option<PrefixBound<'a, K::Prefix>>,
max: Option<PrefixBound<'a, K::Prefix>>,
order: Order
) -> Box<dyn Iterator<Item = StdResult<(K::Output, T)>> + 'c>where
T: 'c,
'a: 'c,
K: 'c,
K::Output: 'static,

While range over a prefix fixes the prefix to one element and iterates over the remaining, prefix_range accepts bounds for the lowest and highest elements of the Prefix itself, and iterates over those (inclusively or exclusively, depending on PrefixBound). There are some issues that distinguish these two, and blindly casting to Vec<u8> doesn’t solve them.

pub fn range_raw<'c>(
&self,
store: &'c dyn Storage,
min: Option<Bound<'a, K>>,
max: Option<Bound<'a, K>>,
order: Order
) -> Box<dyn Iterator<Item = StdResult<Record<T>>> + 'c>where
T: 'c,

pub fn keys_raw<'c>(
&self,
store: &'c dyn Storage,
min: Option<Bound<'a, K>>,
max: Option<Bound<'a, K>>,
order: Order
) -> Box<dyn Iterator<Item = Vec<u8>> + 'c>

pub fn range<'c>(
&self,
store: &'c dyn Storage,
min: Option<Bound<'a, K>>,
max: Option<Bound<'a, K>>,
order: Order
) -> Box<dyn Iterator<Item = StdResult<(K::Output, T)>> + 'c>where
T: 'c,
K::Output: 'static,

pub fn keys<'c>(
&self,
store: &'c dyn Storage,
min: Option<Bound<'a, K>>,
max: Option<Bound<'a, K>>,
order: Order
) -> Box<dyn Iterator<Item = StdResult<K::Output>> + 'c>where
T: 'c,
K::Output: 'static,